find-your-teammate-api/requirements.txt
argcomplete==1.12.3
boto3==1.18.26
botocore==1.21.26
certifi==2021.5.30
cfn-flip==1.2.3
charset-normalizer==2.0.4
click==8.0.1
dnspython==1.16.0
durationpy==0.5
Flask==2.0.0
Flask-PyMongo==2.3.0
future==0.18.2
hjson==3.0.2
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==2.0.1
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
PyJWT==2.1.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tqdm==4.62.1
troposphere==2.7.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wsgi-request-logger==0.4.6
zipp==3.5.0
